+++ CHANGELOG.md
@@ -8,8 +8,12 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 * `Plutarch.Extra.Bool` : `pcond` function for lisp-like boolean 
   conditional-chaining.
-* `Plutarch.Extra.Value` : `phasOnlyOneTokenOfAssetClass` function for checking 
-  that a `PValue` contains exactly one of a `PAssetClass`
+* `Plutarch.Extra.Value` : 
+  * `phasOneTokenOfAssetClass` function for checking 
+  that a `PValue` contains exactly one of a `PAssetClass`.
+  * `phasOneTokenOfSymbol` function for checking that a `PValue` contains exactly
+  one token of an `AssetClass`.
+  
 
 ## 3.9.0 -- 2022-09-23